:shirt: Move test framework classes
