+++ src/SceneGate.Games.ProfessorLayton.Tests/Containers/Binary2ContainerTests.cs
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 SceneGate
+﻿// Copyright (c) 2021 SceneGate
 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -17,94 +17,95 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
+namespace SceneGate.Games.ProfessorLayton.Tests.Containers;
+
 using FluentAssertions;
 using NUnit.Framework;
+using Yarhl.Experimental.TestFramework;
+using Yarhl.Experimental.TestFramework.FluentAssertions;
 using Yarhl.FileFormat;
 using Yarhl.FileSystem;
 using Yarhl.IO;
 
-namespace SceneGate.Games.ProfessorLayton.Tests.Containers
+public abstract class Binary2ContainerTests
 {
-    public abstract class Binary2ContainerTests
+    private int initialStreams;
+    private BinaryFormat original;
+    private NodeContainerInfo containerInfo;
+    private IConverter<BinaryFormat, NodeContainerFormat> containerConverter;
+    private IConverter<NodeContainerFormat, BinaryFormat> binaryConverter;
+
+    [OneTimeSetUp]
+    public void SetUpFixture()
     {
-        private int initialStreams;
-        private BinaryFormat original;
-        private NodeContainerInfo containerInfo;
-        private IConverter<BinaryFormat, NodeContainerFormat> containerConverter;
-        private IConverter<NodeContainerFormat, BinaryFormat> binaryConverter;
-
-        [OneTimeSetUp]
-        public void SetUpFixture()
-        {
-            containerInfo = GetContainerInfo();
-            containerConverter = GetToContainerConverter();
-            binaryConverter = GetToBinaryConverter();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            original?.Dispose();
-
-            // Make sure we didn't leave anything without dispose.
-            Assert.That(DataStream.ActiveStreams, Is.EqualTo(initialStreams), "Missing stream disposes");
-        }
-
-        [SetUp]
-        public void SetUp()
-        {
-            // By opening and disposing in each we prevent other tests failing
-            // because the file is still open.
-            initialStreams = DataStream.ActiveStreams;
-            original = GetBinary();
-        }
-
-        [Test]
-        public void NullToContainerThrowsException()
-        {
-            Assert.That(() => containerConverter.Convert(null), Throws.ArgumentNullException);
-        }
-
-        [Test]
-        public void NullToBinaryThrowsException()
-        {
-            if (binaryConverter == null) {
-                Assert.Ignore();
-            }
-
-            Assert.That(() => binaryConverter.Convert(null), Throws.ArgumentNullException);
-        }
-
-        [Test]
-        public void TransformToContainer()
-        {
-            // Check nodes are expected
-            using var nodes = containerConverter.Convert(original);
-            nodes.Root.Should().MatchInfo(containerInfo);
-
-            // Check everything is virtual node (only the binary stream)
-            DataStream.ActiveStreams.Should().Be(initialStreams + 1);
-        }
-
-        [Test]
-        public void TransformBothWays()
-        {
-            if (binaryConverter == null) {
-                Assert.Ignore();
-            }
-
-            using var nodes = containerConverter.Convert(original);
-            using var actualBinary = binaryConverter.Convert(nodes);
-
-            Assert.That(original.Stream.Compare(actualBinary.Stream), Is.True, "Streams are not identical");
-        }
-
-        protected abstract BinaryFormat GetBinary();
-
-        protected abstract NodeContainerInfo GetContainerInfo();
-
-        protected abstract IConverter<BinaryFormat, NodeContainerFormat> GetToContainerConverter();
-
-        protected abstract IConverter<NodeContainerFormat, BinaryFormat> GetToBinaryConverter();
+        containerInfo = GetContainerInfo();
+        containerConverter = GetToContainerConverter();
+        binaryConverter = GetToBinaryConverter();
     }
+
+    [TearDown]
+    public void TearDown()
+    {
+        original?.Dispose();
+
+        // Make sure we didn't leave anything without dispose.
+        Assert.That(DataStream.ActiveStreams, Is.EqualTo(initialStreams), "Missing stream disposes");
+    }
+
+    [SetUp]
+    public void SetUp()
+    {
+        // By opening and disposing in each we prevent other tests failing
+        // because the file is still open.
+        initialStreams = DataStream.ActiveStreams;
+        original = GetBinary();
+    }
+
+    [Test]
+    public void NullToContainerThrowsException()
+    {
+        Assert.That(() => containerConverter.Convert(null), Throws.ArgumentNullException);
+    }
+
+    [Test]
+    public void NullToBinaryThrowsException()
+    {
+        if (binaryConverter == null) {
+            Assert.Ignore();
+        }
+
+        Assert.That(() => binaryConverter.Convert(null), Throws.ArgumentNullException);
+    }
+
+    [Test]
+    public void TransformToContainer()
+    {
+        // Check nodes are expected
+        using var nodes = containerConverter.Convert(original);
+        nodes.Root.Should().MatchInfo(containerInfo);
+
+        // Check everything is virtual node (only the binary stream)
+        DataStream.ActiveStreams.Should().Be(initialStreams + 1);
+    }
+
+    [Test]
+    public void TransformBothWays()
+    {
+        if (binaryConverter == null) {
+            Assert.Ignore();
+        }
+
+        using var nodes = containerConverter.Convert(original);
+        using var actualBinary = binaryConverter.Convert(nodes);
+
+        Assert.That(original.Stream.Compare(actualBinary.Stream), Is.True, "Streams are not identical");
+    }
+
+    protected abstract BinaryFormat GetBinary();
+
+    protected abstract NodeContainerInfo GetContainerInfo();
+
+    protected abstract IConverter<BinaryFormat, NodeContainerFormat> GetToContainerConverter();
+
+    protected abstract IConverter<NodeContainerFormat, BinaryFormat> GetToBinaryConverter();
 }
