+++ src/SceneGate.Games.ProfessorLayton.Tests/Graphics/BinaryNccg2IndexedImageTests.cs
@@ -24,6 +24,8 @@ namespace SceneGate.Games.ProfessorLayton.Tests.Graphics
     using System.Linq;
     using NUnit.Framework;
     using SceneGate.Games.ProfessorLayton.Graphics;
+    using Yarhl.Experimental.TestFramework;
+    using Yarhl.Experimental.TestFramework.FluentAssertions;
     using Texim.Formats;
     using Texim.Palettes;
     using Yarhl.FileSystem;
