+++ src/SceneGate.Games.ProfessorLayton.Tests/Graphics/BinaryNccl2PaletteCollectionTests.cs
@@ -26,7 +26,8 @@ using System.Linq;
 using FluentAssertions;
 using NUnit.Framework;
 using SceneGate.Games.ProfessorLayton.Graphics;
-using SceneGate.Games.ProfessorLayton.Tests;
+using Yarhl.Experimental.TestFramework;
+using Yarhl.Experimental.TestFramework.FluentAssertions;
 using SixLabors.ImageSharp.Formats.Bmp;
 using Texim.Formats;
 using Yarhl.FileSystem;
