+++ src/SceneGate.Games.ProfessorLayton.Tests/TestFramework/NodeContainerInfo.cs
@@ -17,51 +17,50 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace SceneGate.Games.ProfessorLayton.Tests;
+namespace Yarhl.Experimental.TestFramework;
 
-using System;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.IO;
-using System.Security.Cryptography;
+using System.Linq;
 using YamlDotNet.Serialization;
 using YamlDotNet.Serialization.NamingConventions;
-using Yarhl.IO;
+using Yarhl.FileSystem;
 
-public class BinaryInfo
+public class NodeContainerInfo
 {
-    public long Offset { get; set; }
+    public string Name { get; set; }
 
-    public long Length { get; set; }
+    public string FormatType { get; set; }
 
-    public string Sha256 { get; set; }
+    public BinaryInfo Stream { get; set; }
 
-    public static BinaryInfo FromYaml(string path)
+    public Dictionary<string, object> Tags { get; set; }
+
+    public bool CheckChildren { get; set; }
+
+    public Collection<NodeContainerInfo> Children { get; set; }
+
+    public static NodeContainerInfo FromYaml(string path)
     {
         string yaml = File.ReadAllText(path);
         return new DeserializerBuilder()
             .WithNamingConvention(UnderscoredNamingConvention.Instance)
             .Build()
-            .Deserialize<BinaryInfo>(yaml);
+            .Deserialize<NodeContainerInfo>(yaml);
     }
 
-    public static BinaryInfo FromStream(Stream stream)
+    public static NodeContainerInfo FromNode(Node node)
     {
-        if (stream is null) {
-            return null;
-        }
-
-        stream.Position = 0;
-        using var sha256 = SHA256.Create();
-        sha256.ComputeHash(stream);
-        string hash = BitConverter.ToString(sha256.Hash)
-            .Replace("-", string.Empty)
-            .ToLowerInvariant();
-
-        long offset = stream is DataStream dataStream ? dataStream.Offset : 0;
-
-        return new BinaryInfo {
-            Length = stream.Length,
-            Offset = offset,
-            Sha256 = hash,
+        return new NodeContainerInfo {
+            Name = node.Name,
+            FormatType = node.Format?.GetType().FullName,
+            Tags = node.Tags
+                .Where(x => x.Value?.GetType().IsPrimitive || x.Value is string)
+                .ToDictionary(x => x.Key, x => x.Value),
+            Stream = BinaryInfo.FromStream(node.Stream),
+            CheckChildren = true,
+            Children = new(node.Children.Select(FromNode).ToArray()),
         };
     }
